🏗️ Rename `src/views/` folder to `src/pages/`
`pages` is a more conventional name for such a folder in the React ecosystem.
In the new feature-based architecture, each feature will have its own `pages/` folder, so the root ones must be named accordingly.
