+++ src/pages/ActivityFeed/style.scss
