+++ src/routes/ErrorPageHandler.jsx
@@ -3,10 +3,10 @@ import PropTypes from 'prop-types'
 import { useDispatch, useSelector } from 'react-redux'
 import { useLocation } from 'react-router-dom'
 import { getErrorPageCode } from '../redux/selectors/errorPage'
-import { Error401 } from '../views/Error401'
-import { Error403 } from '../views/Error403'
-import { Error404 } from '../views/Error404'
-import { Error500 } from '../views/Error500'
+import { Error401 } from '../pages/Error401'
+import { Error403 } from '../pages/Error403'
+import { Error404 } from '../pages/Error404'
+import { Error500 } from '../pages/Error500'
 import { setErrorPage } from '../redux/reducers/errorPage'
 
 ErrorPageHandler.propTypes = {
