+++ src/routes/Router.jsx
@@ -1,14 +1,14 @@
 import { Routes, Route } from 'react-router-dom'
 import { useSelector } from 'react-redux'
 import { getUserOrganizationId } from '../redux/selectors/user'
-import { UserProfile } from '../views/UserProfile'
-import { Home } from '../views/Home'
-import { OrganizationCreation } from '../views/OrganizationCreation'
-import { Administration } from '../views/Administration'
-import { ProfileSettings } from '../views/ProfileSettings'
-import { Contact } from '../views/Contact'
-import { SignUp } from '../views/SignUp'
-import { ActivityFeed } from '../views/ActivityFeed'
+import { UserProfile } from '../pages/UserProfile'
+import { Home } from '../pages/Home'
+import { OrganizationCreation } from '../pages/OrganizationCreation'
+import { Administration } from '../pages/Administration'
+import { ProfileSettings } from '../pages/ProfileSettings'
+import { Contact } from '../pages/Contact'
+import { SignUp } from '../pages/SignUp'
+import { ActivityFeed } from '../pages/ActivityFeed'
 import { AdminRoute } from './ConditionalRoute/AdminRoute'
 import { GuestRoute } from './ConditionalRoute/GuestRoute'
 import { AuthenticatedRoute } from './ConditionalRoute/AuthenticatedRoute'
